--- frontend/src/features/file/delete/model/useDeleteFileModel.ts
@@ -1,7 +1,0 @@
-
-export const useDeleteFile = () => {
-
-    return {
-        
-    }
-}
