+++ frontend/src/features/file/remove/index.ts
@@ -1,0 +1,1 @@
+export * from "./model/useRemoveFile";
