+++ frontend/src/features/file/remove/model/useRemoveFile.ts
@@ -1,0 +1,20 @@
+"use client";
+
+import { useFileStore } from "@/entities/file";
+import { useTabStore } from "@/entities/tabs";
+
+export const useRemoveFileModel = () => {
+
+    const { removeFile } = useFileStore();
+    const { activePath,  closeTab } = useTabStore();
+
+    const onRemoveFileHandler = (path: string) => {
+        if (activePath == path) closeTab(path);
+        
+        removeFile(path);
+    }
+
+    return {
+        onRemoveFileHandler
+    }
+}
