+++ frontend/src/features/tabs/index.ts
@@ -1,0 +1,2 @@
+export * from "./model/useContextMenuTabs";
+export * from "./ui/TabsContextMenu";
