+++ frontend/src/shared/ui/ContextMenu/index.ts
@@ -1,0 +1,3 @@
+export * from "./ui/AnimatedContextMenu";
+export * from "./model/types";
+export * from "./model/useContextMenu";
