+++ frontend/src/shared/ui/ContextMenu/model/types.ts
@@ -1,0 +1,24 @@
+
+export interface ContextMenuPosition<T> {
+    x: number;
+    y: number;
+    data: T;
+}
+
+export interface ContextMenuItem {
+  label: string;
+  action: () => void;
+  icon?: React.ReactNode;
+  danger?: boolean;
+  disabled?: boolean;
+}
+
+export interface AnimatedContextMenuProps {
+  visible: boolean;
+  x: number;
+  y: number;
+  items: ContextMenuItem[];
+  menuRef: React.RefObject<HTMLElement>;
+  onClose: () => void;
+}
+
